requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
feedparser==6.0.11
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
MetaTrader5==5.0.4424
multidict==6.1.0
npm==0.1.1
numpy==1.26.4
openai==0.28.0
optional-django==0.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
pillow==10.4.0
plotly==5.24.1
polygon==1.2.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pygame==2.6.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.5
PythonMetaTrader5==1.0.9
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.24.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.4
sqlparse==0.5.1
srt==3.5.3
TA-Lib==0.4.32
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vosk==0.3.45
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.11.1
